Publish colored pointcloud as filtered_cloud from OctomapUpdater
